+++ date.h
@@ -21,6 +21,8 @@
  * Boston, MA 02111-1307, USA
 \* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
+#include <cstdint>
+
 struct DateTime
 	{
 	int year;
@@ -42,7 +44,7 @@ struct DateTime
 		{ plus(0, 0, days, 0, 0, 0, 0); }
 	void plus(int y, int mo, int d, int h, int mi, int s, int ms);
 	int minus_days(DateTime& dt2);
-	long long minus_milliseconds(DateTime& dt2);
+	int64_t minus_milliseconds(DateTime& dt2);
 	};
 
 bool operator==(const DateTime&, const DateTime&);
